Fix the pedestal subtraction.
The pedestal was calculated entirely incorrectly, but was compensated for
by the baseline removal.  This fixes the pedestal calculation in the raw
calibration, and doesn't change the baseline removal.

+++ src/TPulseCalib.hxx
@@ -25,7 +25,7 @@ public:
     CP::TCalibPulseDigit* operator()(const CP::TDigitProxy& digit);
 
     /// Get the last pedestal value.
-    int GetPedestal() {return fPedestal;}
+    double GetPedestal() {return fPedestal;}
 
     /// Get the most recent average value.
     double GetAverage() {return fAverage;}
@@ -39,7 +39,7 @@ public:
 private:
 
     /// The most recent pedestal value
-    int fPedestal;
+    double fPedestal;
 
     /// The most recent average channel value
     double fAverage;
